support a `shade-time' setting rather than hard-coding 0.3s
This is a little bit hacky; because of the way extensions are
installed, we need to support both a user installation (where
the schema directory is relative to the extension) and a global
one (where the schema directory is relative to /usr/share, or
more accurately the XDG_DATA_DIRS environment variable).

We handle lookup of the schema ourselves; in order to set the
preference, a user can do

  XDG_DATA_DIRS=${XDG_DATA_DIRS}:${HOME}/.local/share/gnome-shell/extensions/shade-inactive-windows@hepaajan.iki.fi/data dconf-editor

at their shell.